requirements.txt
paho-mqtt==1.6.1
requests==2.22.0
redis~=4.0.2
fastapi~=0.71.0
PyMySQL~=1.0.2
DBUtils~=3.0.1
uvicorn~=0.16.0
pydantic~=1.9.0
SQLAlchemy~=1.4.29
uuid~=1.30
minio~=7.1.2
starlette~=0.17.1
pika~=1.2.0
